Removed the option to use floats as gl color component type. Made some of the color component parameters/operators stricter than before (floats are no longer accepted to reduce the amount of bugs).

+++ SpehsEngine/ApplicationData.h
@@ -50,7 +50,7 @@ namespace spehs
 
 		//Other
 		static int consoleTextSize;
-		static float consoleTextAlpha;
+		static int consoleTextAlpha;
 		static int GUITextSize;
 		static std::string GUITextFontPath;
 		static std::string dataDirectory;//Specified directory where data should be stored. Contains the forward slash suffix
